returns 404 when route is not configured in router
